BIG BOOM DOOM
Damnerz just add a bunch of datas for the project like mutiples sources codes some softwares even the 1.9 Doom SHAREWARE ;) and more... like linuxdoom 1.10 based on the github or the last version of BOOM even dosbox is in this update. Some Homebrews are include in this update to simplify the search process and the setup process for new contributors. There is more but you need search for it. :)